Allow specifying pytorch version for installation
The main motivation for this is that on some computers when installing PyTorch with default options, a very old version (1.8.1) is installed.
The new API and GUI allows specifying a requirement (e.g., ">=1.12" or ">=2").

While this does not address the question of why old version is installed by default, it allows modules to automatically install correct pytorch version and allows users to manually specify a version, if effectively fixes #9